re-write steam launcher to be more manual/generic
